Added rush-parser module and refactored rush-shell::shell::split_arguments() into rush-parser::tokenize().
